IGNITE-26012 Split BinaryTupleComparatorTest (#6310)
